backend/requirements.txt
# ORHUN AI Backend dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.0

# Supabase
supabase==2.9.0
pyjwt[crypto]==2.9.0

# HuggingFace
gradio-client==1.4.0
huggingface-hub==0.25.0

# HTTP
httpx==0.27.0
aiofiles==24.1.0
